Added support for showing incanter plots inline.
The result of the evaluation now has an additional :visual parameter.
When the result of an evaluation is a JFreeChart, the chart is saved to
a png and the link gets sent to the client and is displayed.

Caveat: The incanter.charts namespace fails to load inside the clojail
environment, so currently I use eval instead of the sandbox to execute
the code that comes from the client.